refactor(site): stratify landing/ by role
The landing directory held ~30 files mixing six roles - page sections,
effect components, animation infrastructure, gated wrappers, capability
probes and page composition - with the taxonomy carried only by naming
conventions and the AGENTS.md table. Give each role a home:

  sections/     page sections + their shared bits (tinted-icon,
                specular-cta-preset)
  motion/       GSAP infrastructure + animation hooks (hooks/ merged in)
  effects/      ambient/effect components (antigravity, particle-field,
                hero-aurora, page-background)
  motion-aware/ the capability-gated wrapper family
  gates/        capability policy + probes (unchanged; see previous commit)

Page-level composition stays at the landing/ root as the stable import
surface for the routes. Pure file moves + import rewrites; AGENTS.md
'Things to know' table, hard-rule copy examples and the THIRD_PARTY_NOTICE
paths updated to the new locations. No behavior change.
