+++ apps/site/src/components/landing/motion/split-text-reveal.tsx
