+++ apps/site/src/components/landing/motion/use-viewport-anim.ts
@@ -1,5 +1,5 @@
 import { useEffect, type RefObject } from 'react';
-import { gsap, registerGsapPlugins, ScrollTrigger } from '../gsap-client';
+import { gsap, registerGsapPlugins, ScrollTrigger } from './gsap-client';
 
 interface Target {
   ref: RefObject<HTMLElement | null>;
