+++ apps/site/src/components/landing/sections/terminal-demo.tsx
