+++ apps/site/src/components/landing/sections/terminal-showcase.tsx
@@ -1,7 +1,7 @@
 import { TerminalDemo } from './terminal-demo';
 import { getStrings } from '@/lib/translations';
-import { Reveal } from './reveal';
-import { GsapScaleUp } from './gsap-scale-up';
+import { Reveal } from '../motion/reveal';
+import { GsapScaleUp } from '../motion/gsap-scale-up';
 
 export function TerminalShowcase({ lang }: { lang: string }) {
   const t = getStrings(lang);
